Define a bash prompt
This is for raspberry pi's again, so being a tad hacky with where we are
loading this. The standard .bashrc file for a pi loads .bash_aliases but
nothing else, and I don't want to look after the .bashrc file here.

So since I'm only loading the bash_aliases on raspberry pis, let's load
the prompt here if the file exists